Revert "Database integration - user authentication"
