+++ folio/app/layout.tsx
@@ -5,10 +5,14 @@ export default function RootLayout({
 }: {
   children: React.ReactNode
 }) {
-    return (
-      <html lang="en">
-        <head />
-        <body>{children}</body>
-      </html>
-    )
+  return (
+    <html lang="en">
+      {/*
+        <head /> will contain the components returned by the nearest parent
+        head.tsx. Find out more at https://beta.nextjs.org/docs/api-reference/file-conventions/head
+      */}
+      <head />
+      <body>{children}</body>
+    </html>
+  )
 }
