+++ folio/package.json
@@ -18,10 +18,8 @@
     "eslint-config-next": "13.1.6",
     "grapesjs": "^0.20.4",
     "next": "13.1.6",
-    "pocketbase": "^0.10.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.43.1",
     "react-icons": "^4.7.1",
     "typescript": "4.9.5"
   },
